Add new classes for handling tag related type safety
`TagSafeInteraction` is just a wrapper for an interaction that
raises an exception if the context is not appropiate for tags,
making type checking instantly better.
`TagContext` is a subclass that raises an exception if the
current guild has tags enabled, and implements methods
related to tags but that don't operate directly on these,
as opposed to `Tag` methods.