Split `tags.py` into `tag.py` and `tags.py`
They used to be unified because nothing inside the module was referenced
anywhere else in the project, so it didn't make sense to have a module
that was only gonna be used in one other module. But now `TagContext` is
used in `exceptions.py`, and leaving everything in unified would cause a
circular import that couldn't be fixed without breaking `listeners.py`
due to reloading the exceptions.